+++ ReleaseNotes.md
@@ -21,6 +21,8 @@
 - Added UI element for quest display
 - Added BounceWalk helper, which is a visual effect for low-cost animation
 - Added CombatText system
+- Added "CombatText" action
+- Added automatic combat text for quests, inventory and equipment manager
 - Fixed several UI bugs
 
 ## V1.18.2